back to page generation rather than API
